Move page table utilities to arch-specific header
