+++ src/kernel/diag/mm.c
@@ -1,17 +1,8 @@
 #include "diag/mm.h"
 
+#include "arch/x86_64/pt.h"
 #include "arch/x86_64/registers.h"
 #include "common/libc.h"
-
-/**
- * TODO: move this into a x86_64 memory management header
- */
-#define VM_ADDR_SPACE_SZ 48
-#define VM_MAX_BIT (1lu << (VM_ADDR_SPACE_SZ - 1))
-#define VM_CANON_BITS (~(VM_MAX_BIT - 1))
-#define VM_PG_LV 4
-#define VM_PG_SZ 4096
-#define VM_PG_SZ_BITS 12
 
 static int var_data = 1;
 static int var_bss;
@@ -56,59 +47,6 @@ _check_stk_size(void) {
   int x;
   printf("stksz:  0x%lx\r\n", &x);
   _check_stk_size();
-}
-
-/**
- * Get PML4 page table address.
- */
-static inline void _get_pt_addr(struct cr3_register_pcide *reg_cr3) {
-  __asm__ volatile("mov %%cr3, %%rax\n\t"
-                   "mov %%rax, %0"
-                   : "=m"(*reg_cr3)
-                   :
-                   : "rax");
-}
-
-/**
- * Page-map level X table (levels 2-4).
- *
- * Used for the page directory table (level 2), the page
- * directory pointer table (level 3), and the PML4 (level 4)
- * tables. Not used for the page table (level 1).
- *
- * The avl* fields are ignored by the architecture, and the
- * reserved fields must not be used (set to zero).
- *
- * See IA32/64 Reference, Volume 3A, 4-32
- */
-struct pmlx_entry {
-  uint8_t p : 1;
-  uint8_t rw : 1;
-  uint8_t us : 1;
-  uint8_t pwt : 1;
-  uint8_t pcd : 1;
-  uint8_t a : 1;
-  uint8_t avl : 1;
-  // Reserved on PML4, used to indicate 1GiB/2MiB pages
-  // on PML3 and PML2, respectively.
-  uint8_t ps : 1;
-  uint8_t avl2 : 4;
-  // Assuming a 52-bit physical address space.
-  uint64_t addr : 40;
-  uint16_t avl3 : 11;
-  uint8_t xd : 1;
-} __attribute__((packed));
-
-/**
- * Canonicalize virtual address. For a 48-bit address space
- * (4-level paging) in x86_64, this means sign-extending bit 47
- * through to bit 63.
- */
-void *va_canonicalize(void *addr) {
-  if ((uint64_t)addr & VM_MAX_BIT) {
-    return (void *)((uint64_t)addr | VM_CANON_BITS);
-  }
-  return addr;
 }
 
 void _print_pt(struct pmlx_entry *pmlx, int level, void *prev_va) {
